+++ src/main/java/plugily/projects/buildbattle/arena/ArenaManager.java
@@ -105,12 +105,13 @@ public class ArenaManager {
             if(partyPlayer == player) {
               continue;
             }
-            if(ArenaRegistry.getArena(partyPlayer) != null) {
-              if(ArenaRegistry.getArena(partyPlayer).getArenaState() == ArenaState.IN_GAME) {
+            BaseArena partyPlayerArena = ArenaRegistry.getArena(partyPlayer);
+            if(partyPlayerArena != null) {
+              if(partyPlayerArena.getArenaState() == ArenaState.IN_GAME) {
                 continue;
               }
               Debugger.debug("[Party] Remove party member " + partyPlayer.getName() + " from other not ingame arena " + player.getName());
-              leaveAttempt(partyPlayer, ArenaRegistry.getArena(partyPlayer));
+              leaveAttempt(partyPlayer, partyPlayerArena);
             }
             partyPlayer.sendMessage(chatManager.getPrefix() + chatManager.formatMessage(arena, chatManager.colorMessage("In-Game.Join-As-Party-Member"), partyPlayer));
             joinAttempt(partyPlayer, arena);
@@ -202,7 +203,9 @@ public class ArenaManager {
 
       player.getInventory().setItem(0, new ItemBuilder(XMaterial.COMPASS.parseItem()).name(chatManager.colorMessage("In-Game.Spectator.Spectator-Item-Name")).build());
       player.getInventory().setItem(4, new ItemBuilder(XMaterial.COMPARATOR.parseItem()).name(chatManager.colorMessage("In-Game.Spectator.Settings-Menu.Item-Name")).build());
-      player.getInventory().setItem(8, leaveItem.getItemStack());
+      if (leaveItem != null) {
+        player.getInventory().setItem(8, leaveItem.getItemStack());
+      }
 
       player.getActivePotionEffects().forEach(potionEffect -> player.removePotionEffect(potionEffect.getType()));
       player.addPotionEffect(new PotionEffect(PotionEffectType.NIGHT_VISION, Integer.MAX_VALUE, 0));
@@ -229,7 +232,9 @@ public class ArenaManager {
     arena.addPlayer(player);
 
     arena.teleportToLobby(player);
-    player.getInventory().setItem(leaveItem.getSlot(), leaveItem.getItemStack());
+    if (leaveItem != null) {
+      player.getInventory().setItem(leaveItem.getSlot(), leaveItem.getItemStack());
+    }
     player.updateInventory();
 
     chatManager.broadcastAction(arena, player, ChatManager.ActionType.JOIN);
@@ -248,8 +253,7 @@ public class ArenaManager {
    */
   public static void leaveAttempt(Player player, BaseArena arena) {
     Debugger.debug("Initial leave attempt, " + player.getName());
-    BBGameLeaveEvent bbGameLeaveEvent = new BBGameLeaveEvent(player, arena);
-    Bukkit.getPluginManager().callEvent(bbGameLeaveEvent);
+    Bukkit.getPluginManager().callEvent(new BBGameLeaveEvent(player, arena));
 
     arena.teleportToEndLocation(player);
 
@@ -335,8 +339,7 @@ public class ArenaManager {
    */
   public static void stopGame(boolean quickStop, BaseArena arena) {
     Debugger.debug("Game stop event initiate, arena " + arena.getID());
-    BBGameEndEvent gameEndEvent = new BBGameEndEvent(arena);
-    Bukkit.getPluginManager().callEvent(gameEndEvent);
+    Bukkit.getPluginManager().callEvent(new BBGameEndEvent(arena));
     for(Player player : arena.getPlayers()) {
       if(!quickStop) {
         spawnFireworks(arena, player);