+++ src/main/java/plugily/projects/buildbattle/arena/managers/ScoreboardManager.java
@@ -34,6 +34,7 @@ import plugily.projects.buildbattle.Main;
 import plugily.projects.buildbattle.arena.ArenaState;
 import plugily.projects.buildbattle.arena.impl.BaseArena;
 import plugily.projects.buildbattle.arena.impl.SoloArena;
+import plugily.projects.buildbattle.arena.managers.plots.Plot;
 import plugily.projects.buildbattle.handlers.language.LanguageManager;
 import plugily.projects.buildbattle.handlers.reward.Reward;
 import plugily.projects.buildbattle.user.User;
@@ -129,8 +130,10 @@ public class ScoreboardManager {
     if(arena.getArenaState() == ArenaState.IN_GAME || arena.getArenaState() == ArenaState.ENDING) {
       lines = scoreboardContents.get(arena.getArenaState().getFormattedName() + "_" + arena.getArenaType().getPrefix());
     }
-    for(String line : lines) {
-      builder.next(formatScoreboardLine(line, user));
+    if (lines != null) {
+      for(String line : lines) {
+        builder.next(formatScoreboardLine(line, user));
+      }
     }
     return builder.build();
   }
@@ -140,26 +143,25 @@ public class ScoreboardManager {
     String returnString = string;
     returnString = StringUtils.replace(returnString, "%PLAYERS%", Integer.toString(arena.getPlayers().size()));
     returnString = StringUtils.replace(returnString, "%PLAYER%", player.getName());
+    returnString = replaceValues(returnString);
     if(((SoloArena) arena).isThemeVoteTime()) {
       returnString = StringUtils.replace(returnString, "%THEME%", plugin.getChatManager().colorMessage("In-Game.No-Theme-Yet"));
+      returnString = StringUtils.replace(returnString, "%TEAMMATE%", plugin.getChatManager().colorMessage("In-Game.Nobody"));
     } else {
       returnString = StringUtils.replace(returnString, "%THEME%", arena.getTheme());
-    }
-    returnString = replaceValues(returnString);
-    if(!((SoloArena) arena).isThemeVoteTime()) {
-      if(arena.getArenaType() == BaseArena.ArenaType.TEAM && arena.getPlotManager().getPlot(player) != null) {
-        if(arena.getPlotManager().getPlot(player).getOwners().size() == 2) {
-          if(arena.getPlotManager().getPlot(player).getOwners().get(0).equals(player)) {
-            returnString = StringUtils.replace(returnString, "%TEAMMATE%", arena.getPlotManager().getPlot(player).getOwners().get(1).getName());
+
+      Plot plot = arena.getPlotManager().getPlot(player);
+      if(arena.getArenaType() == BaseArena.ArenaType.TEAM && plot != null) {
+        if(plot.getOwners().size() == 2) {
+          if(plot.getOwners().get(0).equals(player)) {
+            returnString = StringUtils.replace(returnString, "%TEAMMATE%", plot.getOwners().get(1).getName());
           } else {
-            returnString = StringUtils.replace(returnString, "%TEAMMATE%", arena.getPlotManager().getPlot(player).getOwners().get(0).getName());
+            returnString = StringUtils.replace(returnString, "%TEAMMATE%", plot.getOwners().get(0).getName());
           }
         } else {
           returnString = StringUtils.replace(returnString, "%TEAMMATE%", plugin.getChatManager().colorMessage("In-Game.Nobody"));
         }
       }
-    } else {
-      returnString = StringUtils.replace(returnString, "%TEAMMATE%", plugin.getChatManager().colorMessage("In-Game.Nobody"));
     }
     if(plugin.getServer().getPluginManager().isPluginEnabled("PlaceholderAPI")) {
       returnString = PlaceholderAPI.setPlaceholders(player, returnString);