+++ src/main/java/plugily/projects/buildbattle/arena/managers/plots/Plot.java
@@ -39,7 +39,6 @@ import pl.plajerlair.commonsbox.minecraft.dimensional.Cuboid;
 import plugily.projects.buildbattle.Main;
 import plugily.projects.buildbattle.api.event.plot.BBPlotResetEvent;
 import plugily.projects.buildbattle.arena.impl.BaseArena;
-import plugily.projects.buildbattle.user.User;
 import plugily.projects.buildbattle.utils.Utils;
 
 import java.util.ArrayList;
@@ -84,11 +83,6 @@ public class Plot {
 
   public Map<Location, String> getParticles() {
     return particles;
-  }
-
-  @Deprecated
-  public void addParticle(Location location, String effect) {
-    particles.put(location, effect);
   }
 
   public Biome getPlotDefaultBiome() {
@@ -136,8 +130,7 @@ public class Plot {
     resetPlot();
     if(!owners.isEmpty()) {
       for(Player p : owners) {
-        User user = plugin.getUserManager().getUser(p);
-        user.setCurrentPlot(null);
+        plugin.getUserManager().getUser(p).setCurrentPlot(null);
         setOwners(new ArrayList<>());
         setPoints(0);
       }
@@ -164,8 +157,9 @@ public class Plot {
       p.resetPlayerTime();
     }
 
-    if(cuboid.getCenter().getWorld() != null) {
-      for(Entity entity : cuboid.getCenter().getWorld().getEntities()) {
+    Location center = cuboid.getCenter();
+    if(center.getWorld() != null) {
+      for(Entity entity : center.getWorld().getEntities()) {
         if(cuboid.isInWithMarge(entity.getLocation(), 5)) {
           if(plugin.getServer().getPluginManager().isPluginEnabled("Citizens") && CitizensAPI.getNPCRegistry() != null
               && CitizensAPI.getNPCRegistry().isNPC(entity)) {
@@ -197,13 +191,12 @@ public class Plot {
     if(ServerVersion.Version.isCurrentHigher(ServerVersion.Version.v1_15_R1)) {
       int y = Math.min(cuboid.getMinPoint().getBlockY(), cuboid.getMaxPoint().getBlockY());
 
-      cuboid.getCenter().getWorld().setBiome(cuboid.getMinPoint().getBlockX(), y, cuboid.getMaxPoint().getBlockZ(), plotDefaultBiome);
+      center.getWorld().setBiome(cuboid.getMinPoint().getBlockX(), y, cuboid.getMaxPoint().getBlockZ(), plotDefaultBiome);
     } else {
-      cuboid.getCenter().getWorld().setBiome(cuboid.getMinPoint().getBlockX(), cuboid.getMaxPoint().getBlockZ(), plotDefaultBiome);
+      center.getWorld().setBiome(cuboid.getMinPoint().getBlockX(), cuboid.getMaxPoint().getBlockZ(), plotDefaultBiome);
     }
 
-    BBPlotResetEvent event = new BBPlotResetEvent(arena, this);
-    Bukkit.getServer().getPluginManager().callEvent(event);
+    Bukkit.getServer().getPluginManager().callEvent(new BBPlotResetEvent(arena, this));
   }
 
   public int getPoints() {