+++ src/main/java/plugily/projects/buildbattle/utils/ExceptionLogHandler.java
@@ -57,14 +57,14 @@ public class ExceptionLogHandler extends Handler {
   @Override
   public void publish(LogRecord record) {
     Throwable throwable = record.getThrown();
-    if(!(throwable instanceof Exception) || !throwable.getClass().getSimpleName().contains("Exception") || throwable.getStackTrace().length == 0) {
+    if(!(throwable instanceof Exception) || !throwable.getClass().getSimpleName().contains("Exception") || throwable.getCause() == null) {
       return;
     }
-    if(throwable.getCause() != null && (throwable.getCause().getStackTrace().length == 0 ||
-        throwable.getCause().getStackTrace()[0] == null || !throwable.getCause().getStackTrace()[0].getClassName().contains("plugily.projects.buildbattle"))) {
+    StackTraceElement[] element = throwable.getCause().getStackTrace();
+    if(element.length == 0 || element[0] == null || !element[0].getClassName().contains("plugily.projects.buildbattle")) {
       return;
     }
-    if(!throwable.getStackTrace()[0].getClassName().contains("plugily.projects.buildbattle") || containsBlacklistedClass(throwable)) {
+    if(containsBlacklistedClass(throwable)) {
       return;
     }
     new ReportedException(plugin, (Exception) throwable);